biod: Fix error in implementing make_span function
The make_span function expects an iterator and a size. Since iterating a
C struct pointer to the next item would result in an actual address
shift of sizeof(struct), the size should be 1. This was causing a
stack-buffer-overflow when checked using Address Sanitizer.

BUG=b:228882431
TEST=cros_workon_make --board=hatch biod --test
TEST=USE=test emerge-hatch biod

Change-Id: Id072c88ca5b95df941bb003cfeaadb92dee5eda8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4935637
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>