shill: cellular: Override tethering_allowed with chrome flag
Use the flag ExperimentalTetheringFunctionality to allow tethering in
all carriers which don't have the feature specifically disabled.
This is needed to be able to test the feature on new carriers before
enabling the feature on that carrier.
This implementation depends on the assumption that EntitlementCheck is
always called before AcquireTetheringNetwork.

BUG=b:303926457
TEST=Modify the modb to set the tethering_allowed flag to
enabled/disabled/unset and verify that the behavior doesn't change when
the flag is set to enabled/disabled, but it can be overridden when unset

Change-Id: I0fe91dcbba90cfe4a90203cfd414cc440a453fd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4935648
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>