cryptohome: Get rid of UserMetadataReader
This was originally added as an interface for accessing user metadata
from the USS (without loading & decrypting the USS). Now that the USS
interface is been broken up it's not all that useful, acting just as a
very thin wrapper around UssStorage+EncryptedUss. So lets get rid of it.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ibba4e5a7bf69229e9041e8f17d553efcb65820dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4939371
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>