power: Cleanup and fix Adaptive Charging battery state override
This moves the code for overriding battery state for Adaptive Charging
and Charge Limit to where it is otherwise set, making the code easier to
understand.

It also handles several edge cases better, such as only
overriding the state when line power is connected. This helps us clear
the Adaptive Charging and Charge Limited state about 1s earlier. It
extends the upper limit of the display_battery_percentage for overriding
its value as well. This helps avoid the occasional 81% charge showing
up on the UI when Adaptive Charging or Charge Limit are active.

BUG=b:285412932
TEST=tast.power.AdaptiveCharging and unittests

Change-Id: Ifd5a0bc6baa7cd371cef44d699eb2185be689d0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4981710
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Yanye Li <yanyeli@google.com>