chromeos-config: Add min-active-threads to cpu offline preference.
Currently resourced sets the minimum active threads to 2 for
Cpu-offline-half preference. However, this is not known unless
someone reads the resourced code, and it's possible that for some low performance boards, the minimum active threads should be 4 instead of 2.
Add min-active-threads which can be set by the config. If not set, the
default value is 2.

BUG=b:301199299, b:297331144
TEST=./regen.sh

Cq-Depend: chromium:4944031
Change-Id: Ib91d3644f8c968169b0cd97622550068982e65a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4943381
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>