shill: cellular: make the NetworkInfo a unique_ptr
Because std::optional stores the object inline, resetting a
std::optional<NetworkInfo> doesn't guarantee that the optional gets
cleared to std::nullopt before the NetworkInfo destructor gets called.

This is problematic when we disable cellular technology while a
multiplexed DUN connection is up, because the inner Network stop
operation that runs during the destructor gets notified to the
TetheringManager, which in turn starts the process to release the
tethering network, and by the time the cellular object is notified,
we see how the optional is still set, so we try to reset it twice, and
that seems to trigger memory corruption. Several different crashes
were detected, with fairly similar signatures, happening just after
that second reset.

Switching the NetworkInfo objects to be encapsulated within a
std::unique_ptr removes that cycle because the unique pointer is
cleared before the destructor of the NetworkInfo is called. Therefore,
by the time the TetheringManager triggers the tethering network
release we can clearly see that there is no longer a NetworkInfo to
remove.

BUG=b:305685718
TEST=brya/L850/T-Mobile
  # /usr/local/autotest/bin/autologin.py
  # dbus-send --system --fixed --print-reply --dest=org.chromium.flimflam / org.chromium.flimflam.Manager.SetProperty string:TetheringAllowed variant:boolean:true
  # dbus-send --system --fixed --print-reply --dest=org.chromium.flimflam / org.chromium.flimflam.Manager.SetTetheringEnabled boolean:true
  # FF=org.chromium.flimflam && gdbus call --system --dest=$FF -o / -m $FF.Manager.DisableTechnology "'cellular'"

Change-Id: I0272ad29e33510c8841cc3f36762a22c7b11ec7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4949851
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Ujjwal Pande <ujjwalpande@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>