libbrillo: Move compressor classes from dlcservice-metadata
The ZlibCompressor and ZlibDecompressor classes are currently
implemented in dlcservice-metadata. Move these classes to libbrillo so
they can be used by other platform2 projects such as b/300996158.

The main goal of this CL is to move the compressor classes with minimal
code changes for now so it is easier to review. Future CLs are planned
to implement compression and decompression for zlib and gzip formatted
data, as well as adding more comments on how to use the classes.

BUG=b:306036605
TEST=CQ

Cq-Depend: chromium:4953476
Change-Id: I3d4ba96aa0b13d5dd8971c2a844e592094f1b025
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4950235
Reviewed-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Kendrake Tsui <kendraketsui@google.com>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>