crash: anomaly::ParserTest: Distinguish between expected complete text and expected substrings
Right now, "expected_text" is actually an expected substring. This is
confusing. It also makes it impossible to test that the parser stops at
the correct position.

For example, if the input is:
Line before
Line start
Line body
Line end
Line after

And we expect the parser to stop on "Line end", there's no way to test
that. Setting expected_text to
Line start
Line body
Line end
will (confusingly!) not confirm that "Line before" and "Line after" are
not in the text.

Split into an explicit expected_substr for tests that just want to test
a substr and expected_test for tests that have the entire expected
output.

BUG=None
TEST=cros_run_unit_tests --board=eve --packages crash-reporter

Change-Id: I26413a0d1f01549f24b125e2a6148b2a14481d6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4950234
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>