vm_tools: concierge: Switch to array of fds
Dbus requires that all fds it transports are valid, so it is not
possible to return a plain fd from a failable function. Dbus doesn't
have an optional type, but use an array to achieve the same result for
the return value of GetVmMemoryManagementKillsConnection.

BUG=b:254164308
TEST=Boot no-poll build with and w/o flag, observe no crashes

Change-Id: Iebb99dbd0d283dc48a708189b454ee88009fa15b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4950374
Reviewed-by: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>