vm_tools: concierge: Update VM MMS enable flow
Now that the host's VM memory management service client will be
resourced instead of chrome, make concierge the thing that determines
whether or not VmMMS will be active. Doing this in concierge instead of
resourced avoids the need to deal with finch flags in a component that
isn't tied to Chrome's lifecycle.

BUG=b:254164308
TEST=CQ

Change-Id: Ic7ae9b6fd741e817d4828f9121190d46d66d5534
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4950376
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>