shill: Skip portal check on proxy config mode != direct
There exists non-managed networks that can reach the internet only
through proxy. As such, the previous check of only checking for
manageness of the network is not enough.

By skipping portal detection on networks with proxies, it is possible
for non-online networks behind proxy to be reported as online. In the
long term, it is expected for shill's portal detector to be compatible
with web proxies (b/207657239).

BUG=b:302126338
TEST=unit tests
TEST=manually set proxy config to direct;
     observe portal is checked;
     manually set proxy config to WPAD;
     observe portal is not checked;
     manually set proxy config to direct again;
     observe portal is checked again;

Change-Id: I0cfdbbfb85b9ec57b5fdcd66df66d973f0c2ba28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4951298
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Andreea Costinas <acostinas@google.com>