common-mk: Add template to build ebpf object file
The current template in `bpf_gen.gni` only supports building the bpf
skeleton file. For the ebpf usage in patchpanel, we need the object file
directly instead. This patch adds a new template for this purpose.

Modify the generate-ebpf.py script to let it treat the skeleton header
files as optional output, so that the new template can use this script
directly with skipping the skeleton generation step.

BUG=b:296952085
TEST=FEATURES="test" emerge-$BOARD secagentd

Change-Id: I54d20ad45bb29b1564028799d16df7f6bfb20a10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4951018
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Ling <jasonling@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>