cryptohome: Restore the chaps key with full auth
When chaps crash/restart in the background, cryptohome will need to
re-insert the auth data to chaps again.

But the auth data is backed by the user auth, this is a workaround to
restore the chaps key silently when user try to perform a full auth
without notifying the user.

In the long run, we might want to add a new auth intent and pop up an UI
to let the user do the full auth.

BUG=b:273900621, b:302814252
TEST=https://crrev.com/c/4952135
TEST=tast run $DUT hwsec.ChapsRestart

Change-Id: I170feeaef908f6cfa61d59c9432a662073b7cc48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4951335
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Yi Chou <yich@google.com>
Tested-by: Yi Chou <yich@google.com>