shill: Re-register and reconnect on clear CustomApnList
When the CustomApnList is cleared, disconnect, reattach and reconnect to
ensure the old APNs are no longer in use.
Triggering a reattach is not enough since the reattach only
disconnects/connects when the modem dettaches and attaches back.

BUG=b:306215983
TEST=modb has a IA(APN=X) and Default APN(X)
Turn revamp flag ON
Add a Default APN Y and a custom IA APN X
Turn revamp flag off
Verify that a new attach request is triggered, and the DEFAULT APN is X

Change-Id: I7acfbccf069e499733a9df1562786db083a52e23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4953719
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>