rmad: Refactor CrosConfigUtils test
Change the way of creating the cros_config database. Previously the
model name was fixed so we have to create different database according
to the testcase requirement. Now we create a database that satisfies
various scenarios, and use the model name to switch between them.

BUG=b:300209282
TEST=FEATURES=test emerge-dedede rmad

Change-Id: I13d4f2525bfb0426bfce28e5f3ab6c33d93af070
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4954737
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Jeff Lin <jeffulin@google.com>