login: Remove debug logs
The logs were added to investigate tast flakiness.
As it turns out, the notification is received quite often because
there are other processes running and notifying about the exit.
So, the notification received from a different process is not an
error, and neither is the notification received when the subprocess
doesn't exist.

BUG=None
TEST=None

Change-Id: Ic57b370d2befb45812a339cce1e5300171bfd7aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4953363
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>