vm_tools: Metrics for responsive balloon
Add metrics to track balloon size, resize and stall events in the
Virtual Machine Memory Management Service.

Not that the Size10Minutes metric is not actually logged every 10
minutes, it is logged on balloon resize at most every 10 minutes. If the
balloon stayed the same size for a long time, we use SendRepeatedToUMA
to log the metric as if it had been logged every 10 minutes.

Also add the size and resize metrics to the existing
LimitCacheBalloonPolicy so that we can compare with the existing
implementation.

See https://crrev.com/c/4971113 for the metric definitions.

BUG=b:254164308
TEST=CQ, `tast run dut multivm.Lifecycle.arc` and check histograms
manually with and without VMMMS enabled.

Change-Id: Id242dc8ee96c4601dee2cda7ec2dea678cc10320
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4956760
Tested-by: Charles Dick <cwd@google.com>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>