crash: Clean up udev_collector_test
The bluetooth tests and other coredump tests use drivername
as iwlwifi which is logically not correct, in addition
it is assumed that fwdumps are never captured for iwlwifi
driver which won't be always true. Modify kNoCollectDriver
to "disallowed_driver" and add driver name for BT tests.

BUG=b:291344512
TEST=FEATURES=test emerge-<target> crash-reporter passes.

Change-Id: I95685267d01f8683eeb9813b81897d9978fca79a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4957863
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>