cros_healthd: allow fstatfs(64) syscall
When doing dynamic preloads with LD_PRELOAD, our patched version
of glibc will need to verify that the SO is on an executable partition.
This requires an fstatfs (or fstatfs64) call.

We're encountering seccomp failures when enabling the LLVM scudo allocator
as a preloaded library because of the above.

BUG=b:306719120
TEST=CQ

Change-Id: I0ea36b3e2f7af4e625bf3e52e3162f535ac91587
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4961053
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>