fbpreprocessor: support Finch flag
fbpreprocessord queries the value of the CrOSLateBootAllowFirmwareDumps
flag to know if the feature is enabled or not. It also writes that value
to a file on disk. Instead of having to query the Finch flag themselves,
the other processes (crash-reporter, debugd) can simply read the content
of the file.

BUG=b:306785672
TEST=cros-workon-${BOARD} start fbpreprocessord
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord

Cq-Depend: chromium:4973439
Change-Id: I7d6343066f22f2d63948829d7386dff7973854bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4961057
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>