resourced: Use focused field from Reported Chrome Processes
Instead of ignoring any visible tab, only ignore focused tabs reported
by chrome. It is possible for a tab to be visible, but still in the
background in which case it should be treated as only protected.

This fixes an issue where a visible, but not focused tab would take
priority over the focused app during the kill order negotiation.

BUG=b:304679315
TEST=CQ

Change-Id: Ib8e0c7abfaa17d5c29ea0f06c9279e6587792238
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4963793
Auto-Submit: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Vovo Yang <vovoy@chromium.org>
Commit-Queue: Vovo Yang <vovoy@chromium.org>