resourced: schedqos: check uclamp support
Checking sched_util_max is zero is not enough because there is a case
where sched_util_max is originally zero.

Instead of holding the bool value as global variable, cache the bool
value in the context. With this, access to the value has no overhead and
we can remove the dependency to once_cell.

Even if resourced fails to initialize scheduler context, resourced
should keep alive without scheduler feature.

This is tested with a unit test locally. However avoid adding the test
to the source code because we are not sure whether the testing
environment supports uclamp or not.

BUG=b:301035565
TEST=manual

Change-Id: I3a807dd6ea5e490d4d68a681640187ea95c549c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4974938
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Shah Hossain <shahadath@google.com>
Tested-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: Youssef Esmat <youssefesmat@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>