vm_tools: refvm: Add tmpfs /tmp during installation
Sommelier will write a lot of logs if it can't open /tmp for writing,
causing performance issues when logs are being sent to the console.

Mount a tmpfs at /tmp if it is otherwise read-only. This makes sommelier
happy. Don't mount over /tmp if it's writeable to keep the Debian
default.

BUG=b:307126732
TEST=build and install refvm

Change-Id: I7c6a073eb100266b8b13838548a8d081d7b7f2ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4965200
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emil Mikulic <easy@google.com>
Commit-Queue: James Ye <jamesye@google.com>