vm_tools: concierge: Update vmm-swap usage policy
Update vmm-swap usage policy to treat the VM shutting down as disabling
vmm-swap. This is more consistent with the impact on TBW, since
shutting down the VM means we throw away all swapped guest memory and
will need to re-swap everything when the VM is launched again - i.e. the
same impact that disabling vmm-swap normally.

The fact that ActiveAfterEnableDuration is unexpectedly low and
DisableReason indicates that a disproportionate number of VM shutdowns
are disabling vmm-swap suggests that we may be spending a lot of TBW
budget on short login sessions. This change should help avoid that.

For simplicity, this change does not attempt to translate shutdown
requests that may be in the history file into the new scheme - it simply
ignores them.

BUG=b:307193725
TEST=cros_run_unit_tests --board=$BOARD --packages vm_host_tools

Change-Id: Ie4e6c4df98d4ffe438fa6b3a849e7425eb99ae01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4965440
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>