cryptohome: Flatten flatbuffer namespaces
This gets rid of the structure:: and auth_factor:: namespaces for the
cryptohome flatbuffer schema. Originally these namespaces were added to
work around issues with how the flatbuffer binding generation worked but
they're now unnecessary.

Unfortunately this doesn't shorten up the naming in the code as much as
one might hope because a bunch of these names are duplicates of existing
C++ (or protobuf) ones, and so to maintain uniqueness I generally added
a "Serialized" prefix in all of the examples where that was the case. So
the names still end up being a bit bulky.

BUG=b:306672223
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I8fbf301d0d41272685752d099bbcc5552eb5a6e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4967395
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>