init: Drop b/244186883 debugfs/tracefs hack
We no longer support kernel 4.4, and all newer kernels no longer have
the debugfs/tracefs automount bug referred to in these comments.

We do, however, seem to care about the SELinux context for
/sys/kernel/debug/tracing (at least in security.* testing), and after
removing the hack from chromeos_startup, we notice some errors due to
the combination of:

(1) a quirk in our debugfs-init.conf restorecon invocations and
(2) the lack of any automount triggers in many use cases

Details on (1): we never actually run `restorecon` on
/sys/kernel/debug/tracing itself; we run it on the rest of debugfs, and
we run it on tracefs. But before the automount is triggered, the
/sys/kernel/debug/tracing is neither of those things.

Details on (2): in particular, the Tast tests use os.Lstat(), where
lstat(2) purposely doesn't automount. From lstat(2):

       AT_NO_AUTOMOUNT (since Linux 2.6.38)
	      Don't automount the terminal ("basename") component of
	      pathname.  Since Linux 3.1 this flag is ignored.  Since
              Linux 4.11 this flag is implied.

So the combination of (1) and (2) means we're in a nebulous state where
the appropriate SELinux context is not set until some other entity uses
statx(2) (which *does* automount) or otherwise explores the
.../debug/tracing directory.

Side gripe: none of this SELinux stuff actually matters, except as an
exercise in hair loss.

BUG=b:244186883
TEST=check /sys/kernel/tracing/. permissions
TEST=debugd.DRMTraceTool.stable
TEST=security.SELinuxFilesSystem
TEST=`grep '\/sys\/kernel/debug\/tracing' /proc/mounts` still empty at
     boot

Change-Id: I7351a9a74eb7a8e85dc03ac7fc70c31f32624eea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4968604
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Brian Norris <briannorris@chromium.org>