patchpanel: client: Make SetVpnLockdown() async
We have observed that this D-Bus call may cause a deadlock in some
cases (see $BUG for details). Since it's only used in shill now and in
the current use cases, shill doesn't rely on the result/return of this
call for the further operations, we can just change the implementation
to async.

BUG=b:307431944
TEST=FEATURES="test" emerge-$BOARD patchpanel-client shill
TEST=tast run $DUT network.AlwaysOnVPNRouting.strict_mode_*
  network.AlwaysOnVPNRelogin.strict_mode_*; Observed that the tests take
  shorter time (~30s to ~1min) to finish.

Change-Id: I169a57b5fec8bc73d0cd163625726fc451f160d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4970793
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>