diagnostics: Sync mojom for volume button routine
Sync the mojo definition with change CL:4950897. Also update some
out-of-sync comments.

Elements of the union types `RoutineArgument` and `RoutineDetail` are
reordered to match the ordering in the Chromium side. The reordering
should be safe because there is no client for routines other than the
memory routine.

BUG=b:302278435
TEST=FEATURES=test emerge-volteer diagnostics

Change-Id: I3f793d2ff967e6e023c71f346583be1836b6efa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4971394
Reviewed-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>