routing-simulator: Support L3 route and route type
At present, the Route class only supports L2 route. It only handles
output interface and destination prefix. Also, the result of lookup
is output interface though we need the whole information of the matched
route to take records of it and modify the packet content in order to
support netfilter and conntrack in iptables in the future.

In this patch,
- Add support for an L3 route and route type classifications.
  - Handle "default", "via", route type and table id.
- Modify the LookUp function to return a matched route instead of an
output interface.

BUG=b:307460180
TEST=FEATURES="test" emerge-$BOARD routing-simulator

Change-Id: Id75fb19b203b9f44bdd1d9002f512d9fadc4cba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4975621
Tested-by: Namika Sakaban <snamika@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Namika Sakaban <snamika@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>