power: Send Power.DisplayAfterResumeDurationMsOn*
Send a metric of the elapsed time from when the kernel resumed to the
UI is displayed as Power.DisplayAfterResumeDurationMsOn*.
Since this is usually the user noticed DUT comes back from the suspend,
we can use this as user space resuming.
The corresponding update of the histogoram.xml in Chrome is crrev/c/5125492.

BUG=b:256693104
TEST=suspend/resume and check the Power.DisplayAfterResumeDurationMs in chrome://histograms

Cq-Depend: chromium:4964665
Change-Id: If5b6d5bc654792970687cd3f59af752f06428786
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4937654
Auto-Submit: Masami Hiramatsu <mhiramat@google.com>
Commit-Queue: Masami Hiramatsu <mhiramat@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Masami Hiramatsu <mhiramat@google.com>