shill: tethering: accept upstream networks with a portal
This patch changes the validation of the upstream network in
TetheringManager and allows the session to start if a captive portal is
found on the upstream network. Some carrier networks use a captive
portal flow for enrolling the user into a tethering data plan. Since
there is no plan to support that captive portal flow on the tethering
device directly (i.e opening the portal page when the tethering session
starts), allowing the session to start gives a chance to the hotspot
clients to still use the network.

BUG=b:207086661
BUG=b:291845893
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I3fe219d803f9b7427488482a830baa3fdbae5f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4975899
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>