flex_hwis: access server with API key authorization
The hwis server and its OnePlatform APIs are online.
In order to communicate with the server, the service
must obtain authorization from the server through an API
key. To do this, the service needs to include the API key
in the headers of HTTP requests.

In the minijail tool, /run/shill is required to resolve
the host name.

BUG=b:297887750
TEST=emerge-reven chromeos-base/flex_hwis
TEST=cros deploy vm chromeos-base/flex_hwis
TEST=run flex_hwis_tool service with minijail
TEST=API keys can authorize HTTP requests

Change-Id: I95b4a73be12a1194635457797f8e7eb107f88038
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4977017
Tested-by: Ting-Hao Lin <tinghaolin@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>