cryptohome: Use repo relative paths for flatbuffers
This would be a prerequisite for any future improvements to the
generation script in terms of actually auto-generating headers and
providing more robust matching of filenames in the schema.

I tried to do a fix that would also actually make the .bfbs files
contain useful paths but unfortunately couldn't quickly make it work so
this just does the easier task of making the .fbs files themselves and
the generated files be more "full path" friendly.

BUG=b:307380687
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib3f889bfb494d71806fc41361f2a70b8f88d24e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4977783
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>