vm_tools: concierge: Remove Balloon Contention Size Limit
The host client for the VMMMS was changed to kill if anything less than
the requested amount of memory was freed. This was causing occasional
priority inversion if the host requested an inflation at a high priority
that was reduced in size by the contention size limit. Additionally, the
contention size limit was originally added to reduce the chance of
starving ARCVM of memory, but balloon stall detection takes care of that
now.

Thus, remove the balloon contention size limit.

BUG=b:254164308
TEST=CQ, multivm.MemoryCanaryPerf

Change-Id: Ic43a10591e6baef42e3cd3e389811fa1f8589569
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4978265
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Kameron Lutes <kalutes@chromium.org>
Auto-Submit: Kameron Lutes <kalutes@chromium.org>