init: add a ceiling for the clock
Since tlsdated is enforcing a max valid time, integrate one here
too.  This can come up if the RTC is buggy or the battery runs out,
and it loads with a very large value.

We'll clamp the clock to 30 years with the expectation none of this
software will be used that far in the future.  If it is, while we reset
the clock backwards, it might still self-heal with tlsdate, assuming
the code can even still talk to the network.

See the bug for much deeper discussion of tlsdate behavior.

BUG=b:307794426
TEST=CQ passes

Change-Id: Iae087d660b1446a6b63c1467fa666403e414072d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4974477
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>