shill: dns: tune retries for legacy DNS client
This patch is a short term alternative to b/307880493 for improving the
retry behavior of DNS queries during network validation with
PortalDetector.

Metrics indicate that a significant number of web probes done by
PortalDetector are hitting the full 8s timeout of the current DNS
client (~7%). Depending on the number of name servers available, this
can correspond to:
 - one query to one server timing out after 8s.
 - two queries respectively to two servers both timing out after 4s.
 - etc

To improve the reliability and total latency of network validation with
PortalDetector, this patch adds a second retry per name server. In the
worst case, this would lead to the following timeout behavior for a
single web probe:
 - 1 name server:
      - 1st query to A timeout after 4s,
      - 2nd query to A timeout after 4s.
 - 2 name servers A and B
      - 1st query to A timeout after 2s,
      - 1st query to B timeout after 4s,
      - 2nd query to A timeout after 2s.
      - 2nd query to B timeout after 4s.
 - etc

Compared to the current situation, with a single name server this allows
probes that timeout on DNS once to finish 4s earlier.

BUG=b:307880493
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I665c4180583d2789fb27d0dc029f9fce1feee3ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4982889
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>