modemfwd: Map dlcservice install errors to modemfwd errors
dlcservice returns installation errors in 2 ways, one is directly in the
async Install call, and another one is on the GetDlcState call, but on
the second case, the error is inside a proto field. Translate the error
in the proto field into a installation error, so metrics can be reported
correctly, and the error message logged can be interpreted correctly by
the anomaly detector.

BUG=b:307607029
TEST=Added unit test.
Removed the cached DLC from redrix, restart modemfwd and see that the
dlc installation fails and the correct message is logged.

Change-Id: I1f7191d9c4a3bc6bdad4b9bbb219258c054b8c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4985342
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>