shill: Allow disabling the charon rlimit_as
Address space limits are not viable with sanitizers and
large-address space allocators such as scudo. Allow a build
configuration which disables it.

Really this flag should probably be named inherit_address_space_limit
or something similar, as that's what it's actually doing, but we use
'system_wide_scudo' here as that's what's already configured with the
boards.

BUG=b:271214988
TEST=cros build-packages --board=volteer-scudo && \
     cros build-image --board=volteer-scudo -r test && \
     cros flash $dut <image> && \
     labtunnel $dut && \
     tast run $tunnel network.VPNRouting.ikev2_ipv4

Cq-Depend: chromium:4990395
Change-Id: I62fbaf39b6475c52f4289b5cf701f73a2d64c4ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4985344
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>