flex_hwis: send a metric to track the results of data deletion
Send UMA metrics to track the results of data deletion.
If the user does not agree to send data and there is
already device data on the server, flex_hwis should
send a request to the server to delete the hardware
data stored for the device.

chromium CL to define the histogram: crrev.com/c/4985887

BUG=b:297887750
TEST=cros_workon_make --board=reven chromeos-base/flex_hwis --test

Change-Id: I5d1941b52be894dcd0d483c809d860474f4b390e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4986311
Tested-by: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>