routing-simulator: Add RoutingPolicyEntry class
Add RoutingPolicyEntry class with a main interface called
"CreateFromPolicyString" and unit test skeleton files. This class
represents each policy entry in a routing policy table. The main
interface of it is a factory method to create a RoutingPolicyEntry
object from a string in in the form of a policy entry in "ip rule".
The detailed implementation will be added by the follow-up patch.

BUG=b:307460180
TEST=FEATURES="test" emerge-$BOARD routing-simulator

Change-Id: I7e2f8bbdfe63756cbc5010431dc73168ee30b2ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4987103
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Namika Sakaban <snamika@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Namika Sakaban <snamika@google.com>