shill: Trigger immediate scan upon disconnect
An STA can scan for APs while connected to an AP. However, BSS
entries can expire when background scan interval is extended. So
when an STA disconnects from the current AP, the current BSS will
be the only available BSS in shill and supplicant. To repopulate
the BSS list and find an AP to reconnect to, the STA would need to
rescan.

Previously, shill deferred to a timer for the next foreground scan
which might take up to 10 seconds to kick in during which the STA
would be idle. Shill, now, initiates an immediate scan upon
disconnect to handle this edge case and prevent sitting idle for too
long.

The scan triggers the following code path
WiFi::BSSAddedTask --> WiFiProvider::OnEndpointAdded -->
Manager::UpdateService --> Manager::SortServices -->
Manager::SortServicesTask --> AutoConnect() which leads to a
reconnect if the detected service is autoconnectable.

This CL slightly modifies behavior when manually disconnecting
from a service. If there are other services in the cache, the DUT
still reconnects to existing services. However, there is a change in
how fast the user reconnects to existing services: if a user manually
disconnects from a service without triggering a scan from ChromeUI,
there is no longer a 10-second delay before reconnecting to another
service.

BUG=b:302742864
TEST=FEATURES=test emerge-${BOARD} shill
Tests that DUT can reconnects in a timely manner upon AP lost:
TEST=tast run $DUT wifi.RoamAPGone* (this test will have a stricter
timeout in chromium:4990184)

Change-Id: I4c8872cb5d287622d8383500a4d8909438f3d9d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4989585
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Ruth Mekonnen <rmekonnen@google.com>