hiberman: Don't report suspend failures twice
Currently SuspendConductor records SuspendFailure events in two
places which can result in two events to be recorded for a single
failure. Remove the recording from hibernate_inner(), in case of
a failure hibernate_inner() return an error, which causes
hibernate() to record the SuspendFailure.

BUG=b:308483959
TEST=force suspend failure
  chrome://histograms/hibernate.event shows the failure

Change-Id: I59f68a265da4b479be47739d299b4f0977db5e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4990578
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>